Take a pass at applying HTML conversions
This has two failures that I'm currently already aware of.

First, when a translation has no end, the start is added as plaintext
instead of having shorter translations given a chance. This currently
has no ill effect.
However, with a different translation set, it would be possible for this
to break translations - For example, if it recognized --image-link- as a
tag, but the author intended to write <s>image</s><s>link</s> for
some reason, that would break. I can't think of a better example right
now... I might just file this away as an issue to be fixed when I have
nothing better to be implementing.

The second defect is a little bigger - This does not recursively
translate inner text. This prevents us from having nested styling - That
means that if one writes *-text-*, one gets <strong>-text-</strong>,
rather than the perhaps more expected <strong><s>text</s></strong>.
This one, I actually intend to fix. I just want to get this part working
before I refactor it into a more complicated design. Consider this a
testing version - I'll get the page working around it before I make it
work for more complicated texts.

That goes for escapes too. It would be nice to have a way to make a
translation start tag act as plain text - \ works fine for single
characters, but I'm not sure how it'd work for something like --image.