Switch blog 0 to use a PNG form of question-mark
This image is so small that the difference is mediocre, but I'd like to
support fallback anyway on principle. That's not quite there yet though
- I don't know how I'd like to support specification of multiple image
sources in blog posts (syntax-wise), so until then this needs to always
be a PNG.

This was the last image that was always served as WEBP, so this commit
fixes #12.