Update all blog specifications for new syntax
This was a pain for blog 1, but now the source is much more legible and
somewhat more maintainable - For what little maintenance I plan to do -
And the new syntax will be nicer if I ever add articles 3+ or if someone
actually uses the live preview functionality.

This commit fixes #14.