Improve live blog layout on varying sizes
This fixes a number of issues, where the results div is too large and
forces scroll, or is too small and prohibits scroll on short devices.

On most screensizes that I tested, this is pretty much the correct
height that there's no scrolling of the page in either direction, but
there isn't much blank space between the results div and the footer.

The exception is on devices that are very short in height - Like my S8
in landscape mode - Where that layout would force the results div to
effectively zero size. The results div has a minimum size sufficient to
cover a few lines to allow some usability on these devices.