Fix blog item 1 on very narrow screens
Both the subtitle and one of the reference links were too long for the
screen at around 400px width, which caused ugly horizontal scroll and a
bug with sizing the viewport that allowed ContactSection to be scrolled
away (partially).

To solve this, shrink the subtitle and title a little on reasonably
narrow screens (at the same time we're deleting fluff from the header
anyway), and on very narrow 400px screens, allow forced word breaks
anywhere to maintain the screen width.