Remove unused assets
I do want these to exist for posterity, but at this point they've been
in more than a few commits, and this is not the correct place for them
anyway.

Maybe I'll bring them back if I have a chance to media-query them in
place, but that won't make 1.0 (I'm almost certain), so I'm not sure if
that time will ever actually come.