Add some adjustments to small screen contact info
Basically, on my phone the phone number and email address ran into each
other and fucked shit up together, and the intent of this commit is to
fix that problem, provide some margin for appearance's sake, and also
shrink the whole layout on this sort of screen to provide more room on
the screen for the actual content.

This was the last mobile issue I wanted to fix before 1.0 - So this
commit fixes #4.